Inkey Solution Assignment
Mr. Ambani wants to create the guest list for his daughter’s marriage. He doesn’t want to leave out rich people from the city. But he also doesn’t want to call a lot of people.

You, a data scientist at Reliance, are given the task to make sure his invite list is concise yet including most of the rich people in the city.

The problem is you do not have an account for people’s income. Your market research team was able to gather income data for roughly 50,000 people identifying them as rich (Income > $50k/year) or not rich (Income < $50k/year).

Train a model based on this data which will then be used to create the entire invite list. For each person, you have identified 14 attributes based on which you will have to identify them as rich or poor.

The criteria Mr. Ambani has set is ‘roughly 90% rich of all the rich people in the city must be on the invite list’. A winning model would be one which reduces the total invitees satisfying the criteria.